+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # TBD
+* Add parameters to `makeRpcRequest` on Gecko client
+* Add `GetBlockchainStatus` endpoint on Gecko client
+
+# 0.2.0
 * Updated code to use latest Kurtosis version that stops containers
 * Added convenience scripts for rebuilding and running to the `scripts` directory
 * Updated CI check & documentation to reflect the use of the convenience scripts
 * Switched ServiceAvailabilityCheckerCore to use the `health` API's getLiveness call
 * Added `GeckoClient` with some basic endpoint implementations
-* Add parameters to `makeRpcRequest` on Gecko client
-* Add `GetBlockchainStatus` endpoint on Gecko client
+* Enabled staking
+* Added grabbing logs from the controller